#220 apply requests of different teammates
